pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "refinery-cli"
version = "0.1.0"
description = "AI-powered trace analysis and prompt fixes for LangSmith"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "Refinery Team", email = "team@refinery.ai"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pydantic>=2.0",
    "httpx>=0.24.0",
    "click>=8.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "gitpython>=3.1",
    "jinja2>=3.1",
    "openai>=1.0",
    "anthropic>=0.18",
    "langsmith>=0.1",
    "aiofiles>=23.0",
    "structlog>=24.0",
    "pyyaml>=6.0",
    "filelock>=3.12",
    "google-generativeai>=0.5.0",
    "tiktoken>=0.5.0",
    "streamlit>=1.49.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "pre-commit>=3.0",
]

[project.scripts]
refinery = "refinery.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["refinery*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]